+++ src/playlist.h
@@ -1,7 +1,7 @@
 /* 
  * Gmu Music Player
  *
- * Copyright (c) 2006-2012 Johannes Heimansberg (wejp.k.vu)
+ * Copyright (c) 2006-2015 Johannes Heimansberg (wejp.k.vu)
  *
  * File: playlist.h  Created: 060930
  *
@@ -30,14 +30,14 @@ struct _Entry
 	char   filename[256];
 	char   name[PL_ENTRY_NAME_MAX_LENGTH];
 	short  played;
-	int    queue_pos;
+	size_t queue_pos;
 	Entry *next_in_queue;
 };
 
 struct _Playlist
 {
-	int             length;
-	int             played_items;
+	size_t          length;
+	size_t          played_items;
 	PlayMode        play_mode;
 	Entry          *current;
 	Entry          *first, *last;
@@ -56,22 +56,22 @@ int      playlist_add_item(Playlist *pl, const char *file, const char *name);
 int      playlist_add_file(Playlist *pl, const char *filename_with_path);
 int      playlist_insert_item_after(Playlist *pl, Entry *entry, const char *file, const char *name);
 int      playlist_insert_file_after(Playlist *pl, Entry *entry, const char *filename_with_path);
-char    *playlist_get_name(Playlist *pl, int item);
+char    *playlist_get_name(Playlist *pl, size_t item);
 int      playlist_entry_set_name(Entry *entry, const char *name);
-char    *playlist_get_filename(Playlist *pl, int item);
-int      playlist_get_length(Playlist *pl);
+char    *playlist_get_filename(Playlist *pl, size_t item);
+size_t   playlist_get_length(Playlist *pl);
 int      playlist_next(Playlist *pl);
 int      playlist_prev(Playlist *pl);
 int      playlist_set_current(Playlist *pl, Entry *entry);
 Entry   *playlist_get_current(Playlist *pl);
-Entry   *playlist_get_entry(Playlist *pl, int item);
+Entry   *playlist_get_entry(Playlist *pl, size_t item);
 Entry   *playlist_get_first(Playlist *pl);
 Entry   *playlist_get_last(Playlist *pl);
 Entry   *playlist_get_next(Entry *entry);
 Entry   *playlist_get_prev(Entry *entry);
 int      playlist_entry_delete(Playlist *pl, Entry *entry);
 /* Deletes playlist item at position 'item' and returns a reference to the next pl entry */
-Entry   *playlist_item_delete(Playlist *pl, int item);
+Entry   *playlist_item_delete(Playlist *pl, size_t item);
 char    *playlist_get_entry_name(Playlist *pl, Entry *entry);
 char    *playlist_get_entry_filename(Playlist *pl, Entry *entry);
 PlayMode playlist_get_play_mode(Playlist *pl);
@@ -80,9 +80,9 @@ PlayMode playlist_cycle_play_mode(Playlist *pl);
 int      playlist_toggle_random_mode(Playlist *pl);
 void     playlist_reset_random(Playlist *pl);
 int      playlist_get_played(Entry *entry);
-int      playlist_add_dir(Playlist *pl, const char *directory, void (*finished_callback)(int pl_len));
+int      playlist_add_dir(Playlist *pl, const char *directory, void (*finished_callback)(size_t pl_len));
 int      playlist_get_current_position(Playlist *pl);
-int      playlist_entry_get_queue_pos(Entry *entry);
+size_t   playlist_entry_get_queue_pos(Entry *entry);
 int      playlist_entry_enqueue(Playlist *pl, Entry *entry);
 int      playlist_is_recursive_directory_add_in_progress(void);
 #endif
